Defined the driving interface
Our interface has one capability: to calculate the total price given
multiple items; and thus we have treated this as a function. But since
this capability is the driving boundary of our system (meaning that we
do not know what how it is going to be exposed) we treated it as an
abstraction, and gave it its own type to signal that it is a specific
business process.