Defined the driven interface
Relationship of pricing rules to items feels like to be one-to-many, so
defining the rules composed into the items felt a bit misleading and
repetitive. As pricing rules are also linked to their items by the SKU,
we decided to treat them separately.

We do not know how these pricing rules will be fed into our system:
- Is it some hardcoded configuration?
- Is it from a database?
- Is it from a cache?
And that will depend on how many pricing rules our system will need to
support and how frequently.

With all that taken into account, we opted to define getting the pricing
rules as another boundary of our system, one that is driven from our
application. And similar to our driving interface, it is yet another
function with its own type to signal another specific business process.