Refactored test to abstract test base
Since the very system under test is an abstraction and also depends on
another abstraction, that is how the pricing rules are fed into the api;
we refactored the test to also be an abstraction so that we can use it
to test our api on different routes.

For now, as we do not have any specific adapters to both driving and
driven interfaces, we have only one route: that is the checkout api
against an in-memory implementation of getting the pricing rules.

We used the name `Contract` to signal that this group of tests hopefully
exercises the behaviour of the api, so we can leave implementation
details away from tests as they are subject to change a lot more
frequently, breaking tests without offering great value.