Preparatory refactoring on pricing rules
As we approach into supporting a more complex pricing rule, the meal
deal, which is composed by multiple SKUs for that special offer, we
refactored some bits to iteratively progress. Also, we also make sure we
have not altered the behaviour of the api by passing the tests already
in place

We see meal deal pricing rule as another variant of pricing rules. This
refactor brings in the necessary setup to be able to support multiple
variants of pricing rules.

It is important that all variants of pricing rules are of the same type,
so we can flexibly deal with concrete variants that may be later added.
One is going to be added soon, also.

technical detail:
The `sealed` construct makes sure that variants of that type are known
only at compile time, and no other subclasses/implementations may appear
after sealing. We have chosen this construct to attempt to make the
design flexible yet secure.