Added support to discount on meal deals
Currently, we only support a meal deal of 2, meaning that it will
contain a special price for a pair of 2 SKUs. However, given to our
current design, we can support another kind of meal deal that has a
triple of SKUs applicable for its special price, by just expanding on
the variants of the pricing rule. This may seem tedious if we see we
need to support meal deals of n composition of SKUs in the future, and
thus the design needs to be updated to easily expand.